fix(models): break the models/Rekor-client circular import at its root
#1496 moved `ClientTrustConfig`/`SigningConfig`/`TrustedRoot` into
`sigstore/models.py`. `SigningConfig.get_tlogs()` needs `RekorClient`
and `RekorV2Client`, and those clients need `TransparencyLogEntry` to
type their responses -- a genuine circular import at the module level,
worked around with a function-local (lazy) import in `get_tlogs()`.

#1667 tried to fix this by moving the lazy import to the other side of
the cycle (Rekor clients importing `TransparencyLogEntry` lazily
instead), but a maintainer noted it was still just a lazy import in a
different spot, not a real fix.

This resolves the cycle instead of relocating it, by moving the shared
type out of `models.py` into a leaf module:

* `TransparencyLogEntry` now lives in `sigstore/_internal/rekor/entry.py`,
  which only depends on Rekor/merkle/trust internals -- never on
  `sigstore.models` or the Rekor clients. Both `models.py` (via
  `SigningConfig.get_tlogs()`) and the Rekor clients (`client.py`,
  `client_v2.py`) now import it directly, at module scope, with no lazy
  imports anywhere in the graph.
* `InvalidBundle`/`IncompatibleEntry` move to `sigstore/errors.py`
  (alongside the rest of the exception hierarchy), since
  `TransparencyLogEntry` raises them and `errors.py` has no dependency
  on `models.py` either.
* `sigstore.models` re-exports all three via `__all__` for backwards
  compatibility (`from sigstore.models import TransparencyLogEntry` /
  `InvalidBundle` keep working) and so they keep their generated
  API-reference entries under `docs/api/models.md`.

Adds `test/unit/test_imports.py`, which imports every module that used
to sit on the cycle first, in a fresh interpreter, to guard against a
regression, and checks that `SigningConfig.get_tlogs()` no longer needs
a lazy import.

No behavioral or public API change: `sigstore.models.TransparencyLogEntry`,
`InvalidBundle`, and `IncompatibleEntry` are unchanged in name, contents,
and import path.

Fixes #1536
